Moved ImageStorageTable methods to ImageStorageBehavior.
